Embed Program into library to simplify main
