+++ Sources/tunes/Program.swift
@@ -1,16 +1,9 @@
-import ArgumentParser
 import Foundation
 import iTunes
 
 @main
-public struct Program: AsyncParsableCommand {
-  public static let configuration = CommandConfiguration(
-    commandName: "tunes",
-    abstract: "A tool for working with iTunes data.",
-    version: iTunesVersion,
-    subcommands: [BackupCommand.self, PatchCommand.self, RepairCommand.self, BatchCommand.self],
-    defaultSubcommand: BackupCommand.self
-  )
-
-  public init() {}  // This is public and empty to help the compiler.
+public struct Main {
+  static func main() async {
+    await Program.main()
+  }
 }
